Split STM32F7xx_HAL_Driver-integration into two libraries
It seems that objects (like SystemCoreClock) need to be linked before
the main library and functions (like HAL_Delay()) need to be linked
after it. Otherwise either the functions don't override weak definitions
from the library (everything linked before) or the objects are not
linked at all (everything linked after).